requirements.txt
google-cloud-bigquery==3.4.0
google-cloud-storage==2.10.0
Flask==2.2.2
gunicorn==21.2.0
pandas==1.5.1
scikit-learn==1.2.2
pyarrow==10.0.0
pytest==7.0.1
fastapi==0.109.2
uvicorn==0.27.1
db-dtypes==1.2.0
joblib==1.2.0